Airflow/requirements.txt
apache-airflow-providers-google
psycopg2-binary
pandas
requests
numpy
datetime

